raw_audio_files/requirements.txt
# Core ML libraries
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.60.0

# Pyannote dependencies
pyannote.audio>=3.0.0

# SpeechBrain dependencies  
speechbrain>=0.5.0

# Whisper dependencies
openai-whisper>=20230314

# Optional: For better audio format support
librosa>=0.9.0
soundfile>=0.10.0
